[webcanvas][back] OneToMany 제거 후 오류 수정 / EntityGraph 제거 등

+++ webcanvas-service/src/main/java/com/bb/webcanvasservice/domain/game/repository/GameRoomEntranceCustomRepositoryImpl.java
@@ -39,7 +39,7 @@ public class GameRoomEntranceCustomRepositoryImpl implements GameRoomEntranceCus
                         where   gre.gameRoom.id = :gameRoomId
                         and     gre.state in :enteredStates
                     """;
-        return em.createQuery(jpql, Integer.class)
+        return em.createQuery(jpql, Long.class)
                 .setParameter("gameRoomId", gameRoomId)
                 .setParameter("enteredStates", GameRoomEntranceState.entered)
                 .getSingleResult().intValue();
